feat(compiler): represent int64/uint64 as bigint
Map the fixed 64-bit Go integer types int64 and uint64 to TypeScript
bigint so they preserve full-width precision and Go's two's-complement
and modular overflow semantics, even for small values, keeping identity,
equality, JSON, and type-assertion behavior uniform.

Platform-width uint and uintptr stay the number type but carry a runtime
bigint above 2^53 so full 64-bit width survives (for example ^uint(0)).
Their arithmetic routes through the int64/uint64 helper family and
coerces back through a single coerceWideHelperResult owner: unsigned wide
results widen through uint to preserve precision, signed int coerces with
Number. Plain int, the narrower integer types, and reflect.Kind stay
number.

Index, slice-bound, length, and make/channel-size sites coerce
bigint-backed operands with Number. Update the runtime helpers, the gs
overrides and their behavior tests, the regenerated fixtures, and the
design notes to match.

+++ gs/golang.org/x/crypto/cryptobyte/index.ts
@@ -1378,7 +1378,7 @@ function makeTime(
   const millis =
     globalThis.Date.UTC(year, month - 1, day, hour, minute, second) -
     offsetMinutes * 60 * 1000
-  return time.Unix(Math.floor(millis / 1000), 0)
+  return time.Unix(BigInt(Math.floor(millis / 1000)), 0n)
 }
 
 function parseGeneralizedTime(value: string): time.Time | null {
